pallet_staking_async

Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

apply_slash

Manually and permissionlessly applies a deferred slash for a given era.

bond

Take the origin account as a stash and lock up value of its balance. controller will be the account that controls it.

bond_extra

Add some extra amount that have appeared in the stash free_balance into the balance up for staking.

cancel_deferred_slash

Cancels scheduled slashes for a given era before they are applied.

chill

Declare no desire to either validate or nominate.

chill_other

Declare a controller to stop participating as either a validator or nominator.

deprecate_controller_batch

Updates a batch of controller accounts to their corresponding stash account if they are not the same. Ignores any controller accounts that do not exist, and does not operate if the stash and controller are already the same.

force_apply_min_commission

Force a validator to have at least the minimum commission. This will not affect a validator who already has a commission greater than or equal to the minimum. Any account can call this.

force_new_era

Force there to be a new era at the end of the next session. After this, it will be reset to normal (non-forced) behaviour.

force_new_era_always

Force there to be a new era at the end of sessions indefinitely.

force_no_eras

Force there to be no new eras indefinitely.

force_unstake

Force a current staker to become completely unstaked, immediately.

increase_validator_count

Increments the ideal number of validators up to maximum of T::MaxValidatorSet.

kick

Remove the given nominations from the calling validator.

migrate_currency

Migrates permissionlessly a stash from locks to holds.

nominate

Declare the desire to nominate targets for the origin controller.

payout_stakers

Pay out next page of the stakers behind a validator for the given era.

payout_stakers_by_page

Pay out a page of the stakers behind a validator for the given era and page.

prune_era_step

Perform one step of era pruning to prevent PoV size exhaustion from unbounded deletions.

reap_stash

Remove all data structures concerning a staker/stash once it is at a state where it can be considered dust in the staking system. The requirements are:

rebond

Rebond a portion of the stash scheduled to be unlocked.

restore_ledger

Restores the state of a ledger which is in an inconsistent state.

scale_validator_count

Scale up the ideal number of validators by a factor up to maximum of T::MaxValidatorSet.

set_controller

(Re-)sets the controller of a stash to the stash itself. This function previously accepted a controller argument to set the controller to an account other than the stash itself. This functionality has now been removed, now only setting the controller to the stash, if it is not already.

set_invulnerables

Set the validators who cannot be slashed (if any).

set_min_commission

Sets the minimum amount of commission that each validators must maintain.

set_payee

(Re-)set the payment target for a controller.

set_staking_configs

Update the various staking configurations .

set_validator_count

Sets the ideal number of validators.

unbond

Schedule a portion of the stash to be unlocked ready for transfer out after the bond period ends. If this leaves an amount actively bonded less than asset::existential_deposit, then it is increased to the full amount.

update_payee

Migrates an account’s RewardDestination::Controller to RewardDestination::Account(controller).

validate

Declare the desire to validate for the origin controller.

withdraw_unbonded

Remove any stake that has been fully unbonded and is ready for withdrawal.