pallet_staking_async::dispatchables

Function payout_stakers

pub fn payout_stakers<T: Config>(validator_stash: T::AccountId, era: EraIndex)

Pay out next page of the stakers behind a validator for the given era.

validator_stash is the stash account of the validator.

era may be any era between [current_era - history_depth; current_era].

The origin of this call must be Signed. Any account can call this function, even if it is not one of the stakers.

The reward payout could be paged in case there are too many nominators backing the validator_stash. This call will payout unpaid pages in an ascending order. To claim a specific page, use payout_stakers_by_page.`

If all pages are claimed, it returns an error InvalidPage.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::payout_stakers.