pallet_staking_async

Struct Pallet

pub struct Pallet<T>(/* private fields */);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config> Pallet<T>

pub fn ledger( account: StakingAccount<T::AccountId>, ) -> Result<StakingLedger<T>, Error<T>>

Fetches the ledger associated with a controller or stash account, if any.

pub fn payee( account: StakingAccount<T::AccountId>, ) -> Option<RewardDestination<T::AccountId>>

pub fn bonded(stash: &T::AccountId) -> Option<T::AccountId>

Fetches the controller bonded to a stash account, if any.

pub fn slashable_balance_of(stash: &T::AccountId) -> BalanceOf<T>

The total balance that can be slashed from a stash account as of right now.

pub fn slashable_balance_of_vote_weight( stash: &T::AccountId, issuance: BalanceOf<T>, ) -> VoteWeight

Internal impl of Self::slashable_balance_of that returns [VoteWeight].

pub fn weight_of_fn() -> Box<dyn Fn(&T::AccountId) -> VoteWeight>

Returns a closure around slashable_balance_of_vote_weight that can be passed around.

This prevents call sites from repeatedly requesting total_issuance from backend. But it is important to be only used while the total issuance is not changing.

pub fn weight_of(who: &T::AccountId) -> VoteWeight

Same as weight_of_fn, but made for one time use.

pub fn add_era_stakers( current_era: EraIndex, stash: T::AccountId, exposure: Exposure<T::AccountId, BalanceOf<T>>, )

pub fn set_slash_reward_fraction(fraction: Perbill)

pub fn get_npos_targets(bounds: DataProviderBounds) -> Vec<T::AccountId>

Get all the targets associated are eligible for the npos election.

The target snapshot is always single paged.

This function is self-weighing as [DispatchClass::Mandatory].

pub fn do_add_nominator(who: &T::AccountId, nominations: Nominations<T>)

This function will add a nominator to the Nominators storage map, and VoterList.

If the nominator already exists, their nominations will be updated.

NOTE: you must ALWAYS use this function to add nominator or update their targets. Any access to Nominators or VoterList outside of this function is almost certainly wrong.

pub fn do_remove_nominator(who: &T::AccountId) -> bool

This function will remove a nominator from the Nominators storage map, and VoterList.

Returns true if who was removed from Nominators, otherwise false.

NOTE: you must ALWAYS use this function to remove a nominator from the system. Any access to Nominators or VoterList outside of this function is almost certainly wrong.

pub fn do_add_validator(who: &T::AccountId, prefs: ValidatorPrefs)

This function will add a validator to the Validators storage map.

If the validator already exists, their preferences will be updated.

NOTE: you must ALWAYS use this function to add a validator to the system. Any access to Validators or VoterList outside of this function is almost certainly wrong.

pub fn do_remove_validator(who: &T::AccountId) -> bool

This function will remove a validator from the Validators storage map.

Returns true if who was removed from Validators, otherwise false.

NOTE: you must ALWAYS use this function to remove a validator from the system. Any access to Validators or VoterList outside of this function is almost certainly wrong.

pub fn eras_stakers( era: EraIndex, account: &T::AccountId, ) -> Exposure<T::AccountId, BalanceOf<T>>

Returns full exposure of a validator for a given era.

History note: This used to be a getter for old storage item ErasStakers deprecated in v14 and deleted in v17. Since this function is used in the codebase at various places, we kept it as a custom getter that takes care of getting the full exposure of the validator in a backward compatible way.

impl<T: Config> Pallet<T>

pub fn api_nominations_quota(balance: BalanceOf<T>) -> u32

Returns the current nominations quota for nominators.

Used by the runtime API.

pub fn api_eras_stakers( era: EraIndex, account: T::AccountId, ) -> Exposure<T::AccountId, BalanceOf<T>>

pub fn api_eras_stakers_page_count(era: EraIndex, account: T::AccountId) -> Page

pub fn api_pending_rewards(era: EraIndex, account: T::AccountId) -> bool

impl<T: Config> Pallet<T>

pub fn apply_unapplied_slashes(active_era: EraIndex) -> Weight

Apply previously-unapplied slashes on the beginning of a new era, after a delay.

impl<T: Config> Pallet<T>

pub fn bond( origin: OriginFor<T>, value: BalanceOf<T>, payee: RewardDestination<T::AccountId>, ) -> DispatchResult

Take the origin account as a stash and lock up value of its balance. controller will be the account that controls it.

value must be more than the minimum_balance specified by T::Currency.

The dispatch origin for this call must be Signed by the stash account.

Emits Bonded.

NOTE: Two of the storage writes (Self::bonded, Self::payee) are never cleaned unless the origin falls below existential deposit (or equal to 0) and gets removed as dust.

pub fn bond_extra( origin: OriginFor<T>, max_additional: BalanceOf<T>, ) -> DispatchResult

Add some extra amount that have appeared in the stash free_balance into the balance up for staking.

The dispatch origin for this call must be Signed by the stash, not the controller.

Use this if there are additional funds in your stash account that you wish to bond. Unlike bond or unbond this function does not impose any limitation on the amount that can be added.

Emits Bonded.

pub fn unbond( origin: OriginFor<T>, value: BalanceOf<T>, ) -> DispatchResultWithPostInfo

Schedule a portion of the stash to be unlocked ready for transfer out after the bond period ends. If this leaves an amount actively bonded less than asset::existential_deposit, then it is increased to the full amount.

The dispatch origin for this call must be Signed by the controller, not the stash.

Once the unlock period is done, you can call withdraw_unbonded to actually move the funds out of management ready for transfer.

No more than a limited number of unlocking chunks (see MaxUnlockingChunks) can co-exists at the same time. If there are no unlocking chunks slots available Call::withdraw_unbonded is called to remove some of the chunks (if possible).

If a user encounters the InsufficientBond error when calling this extrinsic, they should call chill first in order to free up their bonded funds.

Emits Unbonded.

See also Call::withdraw_unbonded.

pub fn withdraw_unbonded( origin: OriginFor<T>, _num_slashing_spans: u32, ) -> DispatchResultWithPostInfo

Remove any stake that has been fully unbonded and is ready for withdrawal.

Stake is considered fully unbonded once Config::BondingDuration has elapsed since the unbonding was initiated. In rare cases—such as when offences for the unbonded era have been reported but not yet processed—withdrawal is restricted to eras for which all offences have been processed.

The unlocked stake will be returned as free balance in the stash account.

The dispatch origin for this call must be Signed by the controller.

Emits Withdrawn.

See also Call::unbond.

Parameters

num_slashing_spans: Deprecated. Retained only for backward compatibility; this parameter has no effect.

pub fn validate(origin: OriginFor<T>, prefs: ValidatorPrefs) -> DispatchResult

Declare the desire to validate for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

pub fn nominate( origin: OriginFor<T>, targets: Vec<<<T as Config>::Lookup as StaticLookup>::Source>, ) -> DispatchResult

Declare the desire to nominate targets for the origin controller.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

pub fn chill(origin: OriginFor<T>) -> DispatchResult

Declare no desire to either validate or nominate.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

Complexity

Independent of the arguments. Insignificant complexity.

Contains one read.

Writes are limited to the origin account key.

pub fn set_payee( origin: OriginFor<T>, payee: RewardDestination<T::AccountId>, ) -> DispatchResult

(Re-)set the payment target for a controller.

Effects will be felt instantly (as soon as this function is completed successfully).

The dispatch origin for this call must be Signed by the controller, not the stash.

pub fn set_controller(origin: OriginFor<T>) -> DispatchResult

(Re-)sets the controller of a stash to the stash itself. This function previously accepted a controller argument to set the controller to an account other than the stash itself. This functionality has now been removed, now only setting the controller to the stash, if it is not already.

Effects will be felt instantly (as soon as this function is completed successfully).

The dispatch origin for this call must be Signed by the stash, not the controller.

pub fn set_validator_count(origin: OriginFor<T>, new: u32) -> DispatchResult

Sets the ideal number of validators.

The dispatch origin must be Root.

pub fn increase_validator_count( origin: OriginFor<T>, additional: u32, ) -> DispatchResult

Increments the ideal number of validators up to maximum of T::MaxValidatorSet.

The dispatch origin must be Root.

pub fn scale_validator_count( origin: OriginFor<T>, factor: Percent, ) -> DispatchResult

Scale up the ideal number of validators by a factor up to maximum of T::MaxValidatorSet.

The dispatch origin must be Root.

pub fn force_no_eras(origin: OriginFor<T>) -> DispatchResult

Force there to be no new eras indefinitely.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. Thus the election process may be ongoing when this is called. In this case the election will continue until the next era is triggered.

pub fn force_new_era(origin: OriginFor<T>) -> DispatchResult

Force there to be a new era at the end of the next session. After this, it will be reset to normal (non-forced) behaviour.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. If this is called just before a new era is triggered, the election process may not have enough blocks to get a result.

pub fn force_unstake( origin: OriginFor<T>, stash: T::AccountId, _num_slashing_spans: u32, ) -> DispatchResult

Force a current staker to become completely unstaked, immediately.

The dispatch origin must be Root.

Parameters

stash: The stash account to be unstaked.

num_slashing_spans: Deprecated. This parameter is retained for backward compatibility. It no longer has any effect.

pub fn force_new_era_always(origin: OriginFor<T>) -> DispatchResult

Force there to be a new era at the end of sessions indefinitely.

The dispatch origin must be Root.

Warning

The election process starts multiple blocks before the end of the era. If this is called just before a new era is triggered, the election process may not have enough blocks to get a result.

pub fn cancel_deferred_slash( origin: OriginFor<T>, era: EraIndex, validator_slashes: Vec<(T::AccountId, Perbill)>, ) -> DispatchResult

Cancels scheduled slashes for a given era before they are applied.

This function allows T::AdminOrigin to cancel pending slashes for specified validators in a given era. The cancelled slashes are stored and will be checked when applying slashes.

Parameters

era: The staking era for which slashes should be cancelled. This is the era where the slash would be applied, not the era in which the offence was committed.

validator_slashes: A list of validator stash accounts and their slash fractions to be cancelled.

pub fn payout_stakers( origin: OriginFor<T>, validator_stash: T::AccountId, era: EraIndex, ) -> DispatchResultWithPostInfo

Pay out next page of the stakers behind a validator for the given era.

validator_stash is the stash account of the validator.

era may be any era between [current_era - history_depth; current_era].

The origin of this call must be Signed. Any account can call this function, even if it is not one of the stakers.

The reward payout could be paged in case there are too many nominators backing the validator_stash. This call will payout unpaid pages in an ascending order. To claim a specific page, use payout_stakers_by_page.`

If all pages are claimed, it returns an error InvalidPage.

pub fn rebond( origin: OriginFor<T>, value: BalanceOf<T>, ) -> DispatchResultWithPostInfo

Rebond a portion of the stash scheduled to be unlocked.

The dispatch origin must be signed by the controller.

pub fn reap_stash( origin: OriginFor<T>, stash: T::AccountId, _num_slashing_spans: u32, ) -> DispatchResultWithPostInfo

Remove all data structures concerning a staker/stash once it is at a state where it can be considered dust in the staking system. The requirements are:

the total_balance of the stash is below min_chilled_bond or is zero.

or, the ledger.total of the stash is below min_chilled_bond or is zero.

The former can happen in cases like a slash; the latter when a fully unbonded account is still receiving staking rewards in RewardDestination::Staked.

It can be called by anyone, as long as stash meets the above requirements.

Refunds the transaction fees upon successful execution.

Parameters

stash: The stash account to be reaped.

num_slashing_spans: Deprecated. This parameter is retained for backward compatibility. It no longer has any effect.

pub fn kick( origin: OriginFor<T>, who: Vec<<<T as Config>::Lookup as StaticLookup>::Source>, ) -> DispatchResult

Remove the given nominations from the calling validator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed by the controller, not the stash.

who: A list of nominator stash accounts who are nominating this validator which should no longer be nominating this validator.

Note: Making this call only makes sense if you first set the validator preferences to block any further nominations.

pub fn set_staking_configs( origin: OriginFor<T>, min_nominator_bond: ConfigOp<BalanceOf<T>>, min_validator_bond: ConfigOp<BalanceOf<T>>, max_nominator_count: ConfigOp<u32>, max_validator_count: ConfigOp<u32>, chill_threshold: ConfigOp<Percent>, min_commission: ConfigOp<Perbill>, max_staked_rewards: ConfigOp<Percent>, ) -> DispatchResult

Update the various staking configurations .

min_nominator_bond: The minimum active bond needed to be a nominator.

min_validator_bond: The minimum active bond needed to be a validator.

max_nominator_count: The max number of users who can be a nominator at once. When set to None, no limit is enforced.

max_validator_count: The max number of users who can be a validator at once. When set to None, no limit is enforced.

chill_threshold: The ratio of max_nominator_count or max_validator_count which should be filled in order for the chill_other transaction to work.

min_commission: The minimum amount of commission that each validators must maintain. This is checked only upon calling validate. Existing validators are not affected.

RuntimeOrigin must be Root to call this function.

NOTE: Existing nominators and validators will not be affected by this update. to kick people under the new limits, chill_other should be called.

pub fn chill_other(origin: OriginFor<T>, stash: T::AccountId) -> DispatchResult

Declare a controller to stop participating as either a validator or nominator.

Effects will be felt at the beginning of the next era.

The dispatch origin for this call must be Signed, but can be called by anyone.

If the caller is the same as the controller being targeted, then no further checks are enforced, and this function behaves just like chill.

If the caller is different than the controller being targeted, the following conditions must be met:

controller must belong to a nominator who has become non-decodable,

Or:

A ChillThreshold must be set and checked which defines how close to the max nominators or validators we must reach before users can start chilling one-another.

A MaxNominatorCount and MaxValidatorCount must be set which is used to determine how close we are to the threshold.

A MinNominatorBond and MinValidatorBond must be set and checked, which determines if this is a person that should be chilled because they have not met the threshold bond required.

This can be helpful if bond requirements are updated, and we need to remove old users who do not satisfy these requirements.

pub fn force_apply_min_commission( origin: OriginFor<T>, validator_stash: T::AccountId, ) -> DispatchResult

Force a validator to have at least the minimum commission. This will not affect a validator who already has a commission greater than or equal to the minimum. Any account can call this.

pub fn set_min_commission(origin: OriginFor<T>, new: Perbill) -> DispatchResult

Sets the minimum amount of commission that each validators must maintain.

This call has lower privilege requirements than set_staking_config and can be called by the T::AdminOrigin. Root can always call this.

pub fn payout_stakers_by_page( origin: OriginFor<T>, validator_stash: T::AccountId, era: EraIndex, page: Page, ) -> DispatchResultWithPostInfo

Pay out a page of the stakers behind a validator for the given era and page.

validator_stash is the stash account of the validator.

era may be any era between [current_era - history_depth; current_era].

page is the page index of nominators to pay out with value between 0 and num_nominators / T::MaxExposurePageSize.

The origin of this call must be Signed. Any account can call this function, even if it is not one of the stakers.

If a validator has more than Config::MaxExposurePageSize nominators backing them, then the list of nominators is paged, with each page being capped at [Config::MaxExposurePageSize.] If a validator has more than one page of nominators, the call needs to be made for each page separately in order for all the nominators backing a validator to receive the reward. The nominators are not sorted across pages and so it should not be assumed the highest staker would be on the topmost page and vice versa. If rewards are not claimed in Config::HistoryDepth eras, they are lost.

pub fn update_payee( origin: OriginFor<T>, controller: T::AccountId, ) -> DispatchResultWithPostInfo

Migrates an account’s RewardDestination::Controller to RewardDestination::Account(controller).

Effects will be felt instantly (as soon as this function is completed successfully).

This will waive the transaction fee if the payee is successfully migrated.

pub fn deprecate_controller_batch( origin: OriginFor<T>, controllers: BoundedVec<T::AccountId, T::MaxControllersInDeprecationBatch>, ) -> DispatchResultWithPostInfo

Updates a batch of controller accounts to their corresponding stash account if they are not the same. Ignores any controller accounts that do not exist, and does not operate if the stash and controller are already the same.

Effects will be felt instantly (as soon as this function is completed successfully).

The dispatch origin must be T::AdminOrigin.

pub fn restore_ledger( origin: OriginFor<T>, stash: T::AccountId, maybe_controller: Option<T::AccountId>, maybe_total: Option<BalanceOf<T>>, maybe_unlocking: Option<BoundedVec<UnlockChunk<BalanceOf<T>>, T::MaxUnlockingChunks>>, ) -> DispatchResult

Restores the state of a ledger which is in an inconsistent state.

The requirements to restore a ledger are the following:

The stash is bonded; or

The stash is not bonded but it has a staking lock left behind; or

If the stash has an associated ledger and its state is inconsistent; or

If the ledger is not corrupted but its staking lock is out of sync.

The maybe_* input parameters will overwrite the corresponding data and metadata of the ledger associated with the stash. If the input parameters are not set, the ledger will be reset values from on-chain state.

pub fn migrate_currency( origin: OriginFor<T>, stash: T::AccountId, ) -> DispatchResultWithPostInfo

Migrates permissionlessly a stash from locks to holds.

This removes the old lock on the stake and creates a hold on it atomically. If all stake cannot be held, the best effort is made to hold as much as possible. The remaining stake is removed from the ledger.

The fee is waived if the migration is successful.

pub fn apply_slash( origin: OriginFor<T>, slash_era: EraIndex, slash_key: (T::AccountId, Perbill, u32), ) -> DispatchResultWithPostInfo

Manually and permissionlessly applies a deferred slash for a given era.

Normally, slashes are automatically applied shortly after the start of the slash_era. The automatic application of slashes is handled by the pallet’s internal logic, and it tries to apply one slash page per block of the era. If for some reason, one era is not enough for applying all slash pages, the remaining slashes need to be manually (permissionlessly) applied.

For a given era x, if at era x+1, slashes are still unapplied, all withdrawals get blocked, and these need to be manually applied by calling this function. This function exists as a fallback mechanism for this extreme situation, but we never expect to encounter this in normal scenarios.

The parameters for this call can be queried by looking at the UnappliedSlashes storage for eras older than the active era.

Parameters

slash_era: The staking era in which the slash was originally scheduled.

slash_key: A unique identifier for the slash, represented as a tuple:

stash: The stash account of the validator being slashed.

slash_fraction: The fraction of the stake that was slashed.

page_index: The index of the exposure page being processed.

Behavior

The function is permissionless—anyone can call it.

The slash_era must be the current era or a past era. If it is in the future, the call fails with EraNotStarted.

The fee is waived if the slash is successfully applied.

Future Improvement

Implement an off-chain worker (OCW) task to automatically apply slashes when there is unused block space, improving efficiency.

pub fn prune_era_step( origin: OriginFor<T>, era: EraIndex, ) -> DispatchResultWithPostInfo

Perform one step of era pruning to prevent PoV size exhaustion from unbounded deletions.

This extrinsic enables permissionless lazy pruning of era data by performing incremental deletion of storage items. Each call processes a limited number of items based on available block weight to avoid exceeding block limits.

Returns Pays::No when work is performed to incentivize regular maintenance. Anyone can call this to help maintain the chain’s storage health.

The era must be eligible for pruning (older than HistoryDepth + 1). Check EraPruningState storage to see if an era needs pruning before calling.

impl<T: Config> Pallet<T>

pub fn deposit_event(event: Event<T>)