pallet_staking_async

Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

ActiveEra

The active era information, it holds index and start.

Bonded

Map from all locked “stash” accounts to the controller account.

BondedEras

A mapping from still-bonded eras to the first session index of that era.

CanceledSlashPayout

The amount of currency given to reporters of a slash event which was canceled by extraordinary circumstances (e.g. governance).

CancelledSlashes

Cancelled slashes by era and validator with maximum slash fraction to be cancelled.

ChillThreshold

The threshold for when users can start calling chill_other for other validators / nominators. The threshold is compared to the actual number of validators / nominators (CountFor*) in the system compared to the configured max (Max*Count).

ClaimedRewards

History of claimed paged rewards by era and validator.

CurrentEra

The current planned era index.

ElectableStashes

A bounded list of the “electable” stashes that resulted from a successful election.

EraPruningState

Tracks the current step of era pruning process for each era being lazily pruned.

ErasRewardPoints

Rewards for the last Config::HistoryDepth eras. If reward hasn’t been set or has been removed then 0 reward is returned.

ErasStakersOverview

Summary of validator exposure at a given era.

ErasStakersPaged

Paginated exposure of a validator at given era.

ErasTotalStake

The total amount staked for the last Config::HistoryDepth eras. If total hasn’t been set or has been removed then 0 stake is returned.

ErasValidatorPrefs

Exposure of validator at era with the preferences of validators.

ErasValidatorReward

The total validator era payout for the last Config::HistoryDepth eras.

ForceEra

Mode of era forcing.

Invulnerables

Any validators that may never be slashed or forcibly kicked. It’s a Vec since they’re easy to initialize and the performance hit is minimal (we expect no more than four invulnerables) and restricted to testnets.

Ledger

Map from all (unlocked) “controller” accounts to the info regarding the staking.

MaxNominatorsCount

The maximum nominator count before we stop allowing new validators to join.

MaxStakedRewards

Maximum staked rewards, i.e. the percentage of the era inflation that is used for stake rewards. See Era payout.

MaxValidatorsCount

The maximum validator count before we stop allowing new validators to join.

MinCommission

The minimum amount of commission that validators can set.

MinNominatorBond

The minimum active bond to become and maintain the role of a nominator.

MinValidatorBond

The minimum active bond to become and maintain the role of a validator.

MinimumActiveStake

The minimum active nominator stake of the last successful election.

NextElectionPage

Keeps track of an ongoing multi-page election solution request.

Nominators

The map from nominator stash key to their nomination preferences, namely the validators that they wish to support.

OffenceQueue

Stores reported offences in a queue until they are processed in subsequent blocks.

OffenceQueueEras

Tracks the eras that contain offences in OffenceQueue, sorted from earliest to latest.

Payee

Where the reward payment should be made. Keyed by stash.

ProcessingOffence

Tracks the currently processed offence record from the OffenceQueue.

SlashRewardFraction

The percentage of the slash that is distributed to reporters.

UnappliedSlashes

All unapplied slashes that are queued for later.

ValidatorCount

The ideal number of active validators.

ValidatorSlashInEra

All slashing events on validators, mapped by era to the highest slash proportion and slash value of the era.

Validators

The map from (wannabe) validator stash key to the preferences of that validator.

VirtualStakers

Stakers whose funds are managed by other pallets.

VoterSnapshotStatus

Voter snapshot progress status.